move exceptions to public API
